changed suggested reading back to books, added mybooks to books, adjusted css
