+++ src/pages/Home/Home.css
@@ -74,7 +74,6 @@
 }
 
 .intro h1 {
-  font-size: 2rem;
   margin-bottom: 1rem;
 }
